requirements.txt
Flask==2.2.3
biopython==1.81
certifi==2022.12.7
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
networkx==3.1
Werkzeug==2.2.3
gunicorn
regex
pytest
openai
pandas
pymongo
flask_compress
python-dotenv
